+++ data-avro-1_6/src/main/java/com/linkedin/data/avro/AvroAdapter_1_6.java
@@ -16,13 +16,14 @@
 
 package com.linkedin.data.avro;
 
-import com.linkedin.avroutil1.compatibility.AvroCompatibilityHelper;
 import java.io.IOException;
 import java.io.OutputStream;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.io.Decoder;
+import org.apache.avro.io.DecoderFactory;
 import org.apache.avro.io.Encoder;
+import org.apache.avro.io.EncoderFactory;
 
 /**
  * Adapter for Avro 1.6
@@ -31,6 +32,9 @@ import org.apache.avro.io.Encoder;
 @Deprecated
 public class AvroAdapter_1_6 implements AvroAdapter
 {
+  private final DecoderFactory _decoderFactory = DecoderFactory.get();
+  private final EncoderFactory _encoderFactory = EncoderFactory.get();
+
   @Override
   public boolean jsonUnionMemberHasFullName()
   {
@@ -40,36 +44,36 @@ public class AvroAdapter_1_6 implements AvroAdapter
   @Override
   public GenericData.EnumSymbol createEnumSymbol(Schema avroSchema, String enumValue)
   {
-    return AvroCompatibilityHelper.newEnumSymbol(avroSchema, enumValue);
+    return new GenericData.EnumSymbol(avroSchema, enumValue);
   }
 
   @Override
   public Schema stringToAvroSchema(String avroSchemaJson)
   {
-    return Schema.parse(avroSchemaJson);
+    return new Schema.Parser().parse(avroSchemaJson);
   }
 
   @Override
   public Decoder createBinaryDecoder(byte[] bytes) throws IOException
   {
-    return AvroCompatibilityHelper.newBinaryDecoder(bytes);
+    return _decoderFactory.binaryDecoder(bytes, null);
   }
 
   @Override
   public Encoder createBinaryEncoder(OutputStream outputStream) throws IOException
   {
-    return AvroCompatibilityHelper.newBinaryEncoder(outputStream);
+    return _encoderFactory.binaryEncoder(outputStream, null);
   }
 
   @Override
   public Decoder createJsonDecoder(Schema schema, String json) throws IOException
   {
-    return AvroCompatibilityHelper.newJsonDecoder(schema, json);
+    return _decoderFactory.jsonDecoder(schema, json);
   }
 
   @Override
   public Encoder createJsonEncoder(Schema schema, OutputStream outputStream) throws IOException
   {
-    return AvroCompatibilityHelper.newJsonEncoder(schema, outputStream, true);
+    return _encoderFactory.jsonEncoder(schema, outputStream);
   }
 }