+++ CHANGELOG.md
@@ -14,6 +14,15 @@ and what APIs have changed, if applicable.
 
 ## [Unreleased]
 
+## [29.29.1] - 2022-02-10
+- Make DarkGateKeeper configurable for different dark clusters
+
+## [29.29.0] - 2022-02-09
+- Revert avro update introduced in 29.27.0
+
+## [29.28.0] - 2022-02-04
+- Fix weight double-count in D2 SimpleLoadBalancer
+
 ## [29.27.0] - 2022-01-25
 - Update Avro version to 1.92
 
@@ -5173,7 +5182,10 @@ patch operations can re-use these classes for generating patch messages.
 
 ## [0.14.1]
 
-[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.27.0...master
+[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.29.1...master
+[29.29.1]: https://github.com/linkedin/rest.li/compare/v29.29.0...v29.29.1
+[29.29.0]: https://github.com/linkedin/rest.li/compare/v29.28.0...v29.29.0
+[29.28.0]: https://github.com/linkedin/rest.li/compare/v29.27.0...v29.28.0
 [29.27.0]: https://github.com/linkedin/rest.li/compare/v29.26.4...v29.27.0
 [29.26.4]: https://github.com/linkedin/rest.li/compare/v29.26.3...v29.26.4
 [29.26.3]: https://github.com/linkedin/rest.li/compare/v29.26.2...v29.26.3